+++ package.json
@@ -1,7 +1,7 @@
 {
 	"name": "react-hello-webapp",
 	"engines": {
-		"node": "16.x"
+		"node": "18.x"
 	},
 	"version": "1.0.1",
 	"description": "",
