+++ src/front/js/component/navbar.js
@@ -1,26 +1,53 @@
-import React, { useContext, useEffect } from "react";
+import React, { useContext, useEffect, useState } from "react";
 import { Context } from "../store/appContext";
 import { useNavigate, Link } from "react-router-dom";
-import Login from "./login";
-import '../../styles/navbar.css'; // Importa los estilos aquí
+import '../../styles/navbar.css'; 
 
 export const Navbar = () => {
   const navigate = useNavigate();
   const { store, actions } = useContext(Context);
 
+  const [user, setUser] = useState({
+    email: "",
+    password: "",
+    rememberMe: false,
+  });
+
   useEffect(() => {
     actions.getCurrentUser();
-  }, []);
+  }, [actions]);
 
   const handleLogout = () => {
     actions.logout();
     navigate("/");
   };
 
+  const handleChange = (e) => {
+    const { name, value, type, checked } = e.target;
+    setUser((prevUser) => ({
+      ...prevUser,
+      [name]: type === "checkbox" ? checked : value,
+    }));
+  };
+
+  const loginUser = async (e) => {
+    e.preventDefault();
+    const resp = await actions.login(user);
+
+    if (resp) {
+      if (user.rememberMe) {
+        localStorage.setItem("rememberedEmail", user.email);
+      } else {
+        localStorage.removeItem("rememberedEmail");
+      }
+      navigate(store.user.is_admin ? "/admin" : "/");
+    }
+  };
+
   return (
     <nav className="navbar navbar-expand-lg bg-body-tertiary navbarcolor" aria-label="Eleventh navbar example mx-5">
       <div className="container-fluid">
-        {store.user && store.user.email && (
+        {store.user && store.user.email ? (
           <button
             type="button"
             className="btn dos m-2"
@@ -30,9 +57,7 @@ export const Navbar = () => {
           >
             Espacio Novem
           </button>
-        )}
-
-        {!store.user?.email && (
+        ) : (
           <button className="adminbutton">
             Espacio Novem
           </button>
@@ -109,9 +134,90 @@ export const Navbar = () => {
                 </button>
               </li>
             )}
-          </ul>
 
-          {!store.user?.email ? <Login /> : null}
+            {!store.user?.email && (
+              <li className="nav-item">
+                <button
+                  type="button"
+                  className="btn btn-secondary text-light"
+                  data-bs-toggle="modal"
+                  data-bs-target="#loginModal"
+                >
+                  Iniciar Sesión
+                </button>
+              </li>
+            )}
+          </ul>
+        </div>
+      </div>
+
+      <div className="modal fade" id="loginModal" tabIndex="-1" aria-labelledby="loginModalLabel" aria-hidden="true">
+        <div className="modal-dialog">
+          <div className="modal-content">
+            <div className="modal-header">
+              <h5 className="modal-title" id="loginModalLabel">Iniciar Sesión</h5>
+              <button type="button" className="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
+            </div>
+            <div className="modal-body">
+              <form onSubmit={loginUser}>
+                <div className="mb-3">
+                  <label htmlFor="exampleInputEmail1" className="form-label">
+                    Correo Electrónico
+                  </label>
+                  <input
+                    type="email"
+                    name="email"
+                    value={user.email}
+                    className="form-control form-control-lg dos"
+                    id="exampleInputEmail1"
+                    aria-describedby="emailHelp"
+                    onChange={handleChange}
+                  />
+                </div>
+                <div className="mb-3">
+                  <label htmlFor="exampleInputPassword1" className="form-label">
+                    Contraseña
+                  </label>
+                  <input
+                    type="password"
+                    name="password"
+                    value={user.password}
+                    className="form-control form-control-lg dos"
+                    id="exampleInputPassword1"
+                    onChange={handleChange}
+                  />
+                </div>
+                <div className="mb-3 form-check">
+                  <input
+                    type="checkbox"
+                    className="form-check-input"
+                    id="rememberMe"
+                    name="rememberMe"
+                    checked={user.rememberMe}
+                    onChange={handleChange}
+                  />
+                  <label className="form-check-label ms-2 " htmlFor="rememberMe">
+                    Recordar usuario
+                  </label>
+                </div>
+                <div className="mb-3">
+                  <Link to="/register" className="custom-link">
+                    <p>¿No tienes cuenta? Regístrate</p>
+                  </Link>
+                </div>
+                <div className="mb-3">
+                  <Link to="/send-email" className="custom-link">
+                    <p>¿Olvidaste tu contraseña?</p>
+                  </Link>
+                </div>
+                <div className="d-flex justify-content-end">
+                  <button className="btn btn-primary" type="submit">
+                    Ingresar
+                  </button>
+                </div>
+              </form>
+            </div>
+          </div>
         </div>
       </div>
     </nav>
