+++ src/front/styles/login.css
@@ -1,4 +1,4 @@
-
+/* 
 .uno {
   background-color: #F2E5D5;
 }
@@ -56,4 +56,4 @@
 
 .form-control-lg {
   padding: 0.75rem;
-}
+} */
